Beginnings of a default Mixin for getting frontend editing working
For when your ModelAdmin doesn't do anything special, we'll provide a
class that responds to frontend requests nicely. For now, we just know
we're going to be using the `get_registered_modeladmins` to look for
ModelAdmins which have the attribute `frontend_editing` and do no
deeper inspection.

This leaves the way open for other apps to expose different response
functionality in the frontend.